requirements.txt
google-api-python-client
python-dotenv
polars
transformers
torch
streamlit
plotly
textblob